Modification du calcul de la transmittance pour la voxelisation (temporaire). Modification de la gestion des évènements pour la vue 3D.

+++ trunk/AMAPVox/JeebLight/src/main/java/fr/amap/lidar/amapvox/jeeb/workspace/sunrapp/light/Sun.java
@@ -151,7 +151,7 @@ public class Sun {
 		float weightSum = 0;
 		// plane angle corresponding to sector solid angle of 2PI/nb sectors
 		// Note: solid angle W = 2PI (1-cos(angle/2)) => cos(angle/2)= (46-1) / 46
-		float turtleSectorRadius = (float) Math.acos ((turtle.getNbDirections()-1) / turtle.getNbDirections());
+		float turtleSectorRadius = (float) Math.acos ((turtle.getNbDirections()-1) / (float)turtle.getNbDirections());
 		float sunHaloRadius = turtleSectorRadius; // TODO create method setsunHaloRadius
 		sunHaloRadius /= 2;
 		// System.out.println ("turtle radius= "+Math.toDegrees(turtleSectorRadius));
